Fix instruction for remove
The remove command listed in the README.md doesn't work (for me).  I get the following warning:
> [WARN] [run:rm binary] ./dist/bin/hello container=example-hello-dre-binary Failed to remove container: No such container: example-hello-dre-binary

The command I'm suggesting here at least removes the two containers that are built by the example, but it leaves the dist/bin/hello binary in place.  Gotta dig into the code a bit to figure out how to fix that.