Use file libraries to properly filter excludes
After adding testing, we discovered that the current method of filtering
only works if you're filtering top-level files or directories, which
satisfied the use case of skippiing the '.git' directory.

This add testing, shifts to passing search parameters as a struct, and
uses proper pattern matching to evaluate the dockerignore file based
on the Docker documentation.

Signed-off-by: Tom Duffield <tom@chef.io>